requirements/base.txt
docopt==0.6.2
feedparser==5.1.3
Jinja2==2.7.3
lxml==3.3.5
PyYAML==3.11
requests==2.3.0
